Design: separate CreateVoteModal into modal and button
