--- src/components/Vote/CreatVoteTitleModal/CreatVoteTitleModal.module.scss
@@ -1,6 +1,0 @@
-@use "@/sass" as *;
-
-.container {
-  display: flex;
-  margin-top: 48px;
-}
