+++ src/recoil/vote/createVoteTitleModal.ts
@@ -1,0 +1,6 @@
+import {atom} from 'recoil';
+
+export const isCreateModalOpenState = atom<boolean>({
+  key: 'isCreateModalOpenState',
+  default: false,
+});
